backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
huggingface_hub
